requirements.txt
async-generator==1.10
attrs==23.1.0
black==23.3.0
certifi==2023.5.7
click==8.1.3
exceptiongroup==1.1.1
h11==0.14.0
idna==3.4
isort==5.12.0
mypy-extensions==1.0.0
outcome==1.2.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.5.1
PySocks==1.7.1
python-dotenv==1.0.0
selenium==4.9.1
sniffio==1.3.0
sortedcontainers==2.4.0
tomli==2.0.1
trio==0.22.0
trio-websocket==0.10.2
typing_extensions==4.5.0
urllib3==2.0.2
wsproto==1.2.0
